Add link_whole support for shared libraries
Linker may drop symbols from archive if they are not referenced directly.
For shared library linker may drop the entire library if nothing is referenced
from it. This makes link_whole very similar for both cases. It also enables
transparent switching between 'static' and 'shared' for default-library option.
Before this patch if library used in link_whole was created with
default-library=static, when user switches the flag, configuration fails.
Now shared library will be also linked and will not be dropped even if it is
not referenced directly.